+++ test/unit/conftest.py
@@ -1,0 +1,16 @@
+import os
+from unittest import mock
+
+import pytest
+
+
+@pytest.fixture(autouse=True, scope="session")
+def set_config(monkeypatch):
+    with mock.patch.dict(os.environ, clear=True):
+        envvars = {
+            "DOMAIN ": "localhost",
+            "YOUTUBE_API_KEY": "deadbeef",
+        }
+        for k, v in envvars.items():
+            monkeypatch.setenv(k, v)
+        yield
